feat: epv7 eth module
